Switch to velocity controllers to overcome bug from position controllers
as found in https://github.com/ros-simulation/gazebo_ros_pkgs/issues/479

and reported by the warning

The default_robot_hw_sim plugin is using the Joint::SetPosition method without preserving the link velocity.
As a result, gravity will not be simulated correctly for your model.
Please set gazebo_pid parameters, switch to the VelocityJointInterface or EffortJointInterface, or upgrade to Gazebo 9.